Wait for workspace to be updated in Reanalyze tests
